# ENH: Circular statistic optimization