Hi, 
this pull request concerns 3 functions of circular statistics in the `stats` module: 
* `circmean`
* `circvar`
* `circstd`

The method used until now calls `exp(1j*ang)` in order to calculate `sin` and `cos` of `ang`.
However, calculating the `sin` and `cos` is significantly faster (see below) and also requires half the
RAM at peak usage (really import in the case of large arrays).

Here is a comparison between old (`circmean`) and new (`cirmean2`) methods:
```
In [1]: from scipy import stats

In [2]: import numpy as np

In [3]: phases=np.random.randn(1000)                                      

In [4]: timeit stats.circmean(phases)                                     
1000 loops, best of 3: 357 µs per loop

In [5]: timeit stats.circmean2(phases)
10000 loops, best of 3: 220 µs per loop

In [6]: phases=np.random.randn(1e8)                                       

In [7]: timeit stats.circmean(phases)                                     
1 loops, best of 3: 17.3 s per loop

In [8]: timeit stats.circmean2(phases)
1 loops, best of 3: 12.8 s per loop`
```

This is my first pull request, hope that it's not too messed up. Comments are highly appreciated.