@argriffing 2016-01-03
> `R=sqrt(S**2+C**2)`

did you consider hypot for this?