@kleskjr 2016-01-03
yes, considered, after your suggestion. Thanks!