@rgommers 2016-01-07
@kleskjr thanks, this looks fine module my minor comments. The timings look good, and test coverage is already OK. It would be good to squash all commits though, and write the commit message in this form: http://docs.scipy.org/doc/numpy-dev/dev/gitwash/development_workflow.html#writing-the-commit-message