@kleskjr 2016-01-09
The many commits are squashed. Thanks for the feedback @argriffing, @larsmans, @rgommers, I learned a lot!