@larsmans 2016-01-13
I managed to reproduce your timings and better yet, I found that `stats.circmean(np.random.randn(1e7))` now requires 258MB of RAM instead of 486MB. I suspect memory speed was the bottleneck here.